% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cbind_rcpp}
\alias{cbind_rcpp}
\title{cbind_rcpp}
\usage{
cbind_rcpp(A, B)
}
\arguments{
\A -> Matriz nxm

\B -> Vector n tamaño
}
\value{
Unir un data.frame o matrix con un vector.
}
\description{
Tome una secuencia de argumentos de vector, matriz o marco de datos y combínelos por columnas o filas, respectivamente.
}
\examples{
matriz <- matrix(round(runif(6, 1, 8)), nrow = 3, ncol = 3)
vector <- c(1:3)
matriz
vector
set.seed(122)
result <- cbind_rcpp(matriz,vector)
result
}
