\name{evatopi1-package}
\alias{evatopi1-package}
\alias{evatopi1}
\docType{package}
\title{
  Paquete para la aceleracion de procesos con matrices.
}
\description{
  Matrices y mas matrices, con rcpp.
}
\details{
  ------
}
\author{
Vicente Agustin Milic Fuentes

Maintainer: Your Name <your@email.com>
}
\references{
  Mencion T.I.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.
  }
}
