% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{maxMin_cpp}
\alias{maxMin_cpp}
\title{maxMin_cpp}
\usage{
maxMin_cpp(A, B)
}
\arguments{
\A -> Matriz de mxn

\B -> Matriz de nxn
}
\value{
}
\description{
Esta función le permite escalar vectores o un marco de datos completo utilizando el método de escalado máximo-mínimo. Siempre se devuelve un vector numérico.
}
\examples{
A <- matrix(round(runif(n, min = 0, max = 1), 1), nrow = 3, ncol = 3)
B <- matrix(round(runif(n, min = 0, max = 1), 1), nrow = 3, ncol = 3)
diag(A) <- 1
diag(B) <- 1
set.seed(122)
result <- maxMin_cpp(A,B)
result
}
