% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rbind_rcpp}
\alias{rbind_rcpp}
\title{rbind_rcpp}
\usage{
rbind_rcpp(A, B)
}
\arguments{
\A -> Matriz nxm

\B -> Vector tamaño n
}
\value{
Matriz con vector insertados en una nueva fila
}
\description{
Simula respuestas binarias correlacionadas suponiendo que se cumple un modelo preespecificado para las probabilidades marginales.
}
\examples{
matriz <- matrix(round(runif(6, 1, 8)), nrow = 3, ncol = 3)
vector <- c(1:3)
matriz
vector
set.seed(122)
result <- rbind_rcpp(matriz,vector)
result
}
