% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{whichMax}
\alias{whichMax}
\title{whichMax}
\usage{
whichMax(A)
}
\arguments{
\A -> Matriz nxm
}
\value{
devuelde data.frame
}
\description{
Encontrar el o los máximos valores de un data.frame o matrix.
}
\examples{
matriz <- matrix(round(runif(9, 1, 10)), nrow = 3, ncol = 3)
set.seed(123)
matriz
result <- whichMax(matriz)
result
}
