% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{whichMin}
\alias{whichMin}
\title{whichMin}
\usage{
whichMin(A)
}
\arguments{
\A -> Matriz nxm
}
\value{
devuelve un data.frame
}
\description{
Encontrar el o los mínimos valores de un data.frame o matrix.
}
\examples{
matriz <- matrix(round(runif(6, 1, 8)), nrow = 3, ncol = 3)
set.seed(122)
matriz
result <- whichMin(matriz)
result
}
